add to check install

+++ recipes/mysql.sh
@@ -23,16 +23,17 @@ else
     exit;
 fi
 
-sunzi.install "mysql-server-5.5 libmysqlclient18 libmysqlclient-dev"
-query "SET GLOBAL innodb_fast_shutdown = 0"
-sunzi.mute "service mysql stop"
-sunzi.mute "mkdir -p $targetdir"
-sunzi.mute "rm -f /var/lib/mysql/ib_logfile?"
-echo "MySQL innnodb & characterset configuration setup.."
-cnf="mysql_innodb.cnf"
-targetfile=${targetdir}${cnf}
-rm -f $targetfile
-cat >> $targetfile <<EOM
+if ! sunzi.installed "mysql-server"; then
+    sunzi.install "mysql-server-5.5 libmysqlclient18 libmysqlclient-dev"
+    query "SET GLOBAL innodb_fast_shutdown = 0"
+    sunzi.mute "service mysql stop"
+    sunzi.mute "mkdir -p $targetdir"
+    sunzi.mute "rm -f /var/lib/mysql/ib_logfile?"
+    echo "MySQL innnodb & characterset configuration setup.."
+    cnf="mysql_innodb.cnf"
+    targetfile=${targetdir}${cnf}
+    rm -f $targetfile
+    cat >> $targetfile <<EOM
 [mysqld]
 #
 # * InnoDB
@@ -52,10 +53,10 @@ innodb_doublewrite = false
 default-storage-engine=InnoDB
 EOM
 
-cnf="mysql_characterset.cnf"
-targetfile=${targetdir}${cnf}
-rm -f $targetfile
-cat >> $targetfile <<EOM
+    cnf="mysql_characterset.cnf"
+    targetfile=${targetdir}${cnf}
+    rm -f $targetfile
+    cat >> $targetfile <<EOM
 [client]
 default-character-set=utf8
 
@@ -65,19 +66,20 @@ default-character-set=utf8
 [mysqld]
 character-set-server=utf8
 EOM
-sunzi.mute "service mysql start"
-query "show variables like 'innodb_version'"
+    sunzi.mute "service mysql start"
+    query "show variables like 'innodb_version'"
 
-echo "mysql secure setup start"
-echo "remove anonymous user"
-query "DELETE FROM mysql.user WHERE user='';"
-echo "remove remote access user"
-query "DELETE FROM mysql.user WHERE user='root' AND host NOT IN ('localhost','127.0.0.1','::1');"
-echo "drop test database"
-query "DROP DATABASE IF EXISTS test;"
-echo "remove privileges test database"
-query "DELETE FROM mysql.db WHERE db='test' OR db='test\\_%';"
-echo "flush privileges"
-query "FLUSH PRIVILEGES;"
-echo ""
-echo "YOU MUST BE SET MySQL PASSWORD!!"
+    echo "mysql secure setup start"
+    echo "remove anonymous user"
+    query "DELETE FROM mysql.user WHERE user='';"
+    echo "remove remote access user"
+    query "DELETE FROM mysql.user WHERE user='root' AND host NOT IN ('localhost','127.0.0.1','::1');"
+    echo "drop test database"
+    query "DROP DATABASE IF EXISTS test;"
+    echo "remove privileges test database"
+    query "DELETE FROM mysql.db WHERE db='test' OR db='test%';"
+    echo "flush privileges"
+    query "FLUSH PRIVILEGES;"
+    echo ""
+    echo "YOU MUST BE SET MySQL PASSWORD!!"
+fi
